avocado/core/job.py: fix Avocado version info output to log
The recording of Avocado version, when running from a GIT repo has
been broken for a long time (since the SPEC file was renamed).  This
could be a simple fix to just match the new SPEC file named, but I
think we can do better.  Besides this fix, the changes here include:

 - Print a short commit instead (8 chars)
 - Omit branch name, commit is authoritative enough and developers can
   tell the name of a branch if it's really necessary

But, let me say that this is still broken.  The GIT version
information is only printed if the current work directory is inside
the GIT repo directory.  A better (and way more complex) fix would be
one that looks where the avocado script and Python libraries were
loaded from, and check if that location is a GIT repo.  Instead of
doing that, though, it'd make more sense to use a semantic version
approach (and possibly a library).

Signed-off-by: Cleber Rosa <crosa@redhat.com>