avocado/plugins/replay.py: remove misleading TODO comment
The current state of this comment is the result of a mechanical change
replacing `urls` with `references`.

The source for references given on the command line is working as
intended, that is, if test references are given on the command line
they are taken and will override the ones on the replay args.

Signed-off-by: Cleber Rosa <crosa@redhat.com>