test: Log full path to the test source
Currently it might not be easy to determine what version of the test
source was executed (eg. when test-repo is used). Let's log the test_src
entry which is the location of the source file of the test class.

In the future we or users should be able to extend the concept of test
metadata of other useful details like mentioned in

    https://trello.com/c/XvaiS4UV/1063-introduce-test-repo-concept-to-
    make-test-names-more-relevant

for example "test_tainted, test_version (sha), etc, this is why I chose
generic approach by using `self._get_metadata`.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>