loader: Extract the python unittest loader
When the python unittest loader is part of the FileLoader, it detects
INSTRUMENTED and PyUNITTEST tests for the same directory. The problem is
INSTRUMENTED tests are also python unittests and therefor INSTRUMENTED
tests tagged like "disabled" would be then executed as PyUNITTEST, which
is probably not the expected behavior.

Extracting the loader into standalone loader disallows the possibility
of detecting INSTRUMENTED and PyUNITTESTs from one directory (avocado
run example/tests) but it still allows running both via individual files
(or dirs with only INSTRUMENTED and only PyUNITTESTS).

If we decide this is the expected behavior we should squash this into
the previous commit

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>