[RFC] Alternative implementation of `get_data()`
This alternative implementation of `get_data()` does so in a much more
detached form.  It allows for other test types to clearly change the
way data files are stored and accessed.  It also introduces the concept
of named data sources.

In the previous implementation, the types of "data dirs" were hard coded,
while in this approach, they're defined at the class level and it's up
to the `get_data()` implementation to deal with it.  Consequently, the
`get_data()` method gets a second (optional) parameter that allows the
user to specify from which data source the data file should be retrieved.

This is necessary, for example, when we add configurability to sources
of data files, such as expected `stdout`, `stderr` and `output` files.

About the implementation as a "Mix-In" class: while not strictly
necessary, and clearly a major change in the avocado.core.test.Test
class implementation, it makes that feature easier to grasp, maintain
and modify.  It may also set a precedence for the support of fixtures
in Avocado INSTRUMENTED tests.

Signed-off-by: Cleber Rosa <crosa@redhat.com>