selftests/unit/test_test.py: remote flexmock requirement of test_long_name
Basically, this is not about switching to mock instead, because
mocking properties is more challeging with mock, and mocking
a property of an existing instance is, to the best of my knowledge,
not possible.

So, it seemed like a good idea to use a different test class with an
overloaded property that returns the fake filename, and while at it,
separate split that into two other tests.

It should be easier to understand what each block (now a test) is
about.

Signed-off-by: Cleber Rosa <crosa@redhat.com>