Job: make sure a job can be created without an explicit base_logdir
One of the premises of the Job class is that it can be created without
any mandatory information set in the args parameter.  But, because of
the way we're testing it, we always pass a base_logdir.

This test is about making sure the job gets a base_logdir, and
consequently a logdir, even when a base_logdir is not set in the args.
The implementation of this test is based on the fact that there's
already a tmpdir that can be used for playing the role of a base_logdir.

Signed-off-by: Cleber Rosa <crosa@redhat.com>