Job: make sure a job can be created without an explicit logdir
One of the premises of the Job class is that it can be created without
any mandatory information set in the args parameter.  But, because of
the way we're testing it, we always pass a logdir.  This test is about
making sure the job gets a logdir, even when one is not set in the
args.

Signed-off-by: Cleber Rosa <crosa@redhat.com>