Job: do not count on base_logdir when using dry run mode
As in previous commits, job should not count on arguments being
present, such as "base_logdir".  This also needs to be handled
when using dry run mode.

Since the dry run implementation creates a temporary directory
itself, it needs to be cleaned up in addition to the tmpdir
created by setUp.

Signed-off-by: Cleber Rosa <crosa@redhat.com>