Job: distinguish between the base log dir and the job logdir itself
The Job class code uses the same name, logdir, for both the directory
that will hold a given job's results (job-<timestamp>-<shortid>) and
the base directory, where the first one will be created.

This changes the naming so that the base log dir, only present in
the job arguments (the argparse.Namespace instance) will be named
"base_logdir".  The job's result dir, which value is kept in the
job instance itself, is still named "logdir".

Signed-off-by: Cleber Rosa <crosa@redhat.com>