test: log the raw output on stdout/stderr
Python logging system is hard-coded to add a new line character at the
end of each call to emit(). Since the Avocado recording to test
stdout/stderr uses the logging system, this patch creates/uses a custom
FileHandler to be used in the cases where we need the raw output to
be recorded to the files.

Reference: https://trello.com/c/dH7FLT75
Signed-off-by: Amador Pahim <apahim@redhat.com>