selftests/unit/test_jobdata.py: split test into one per version
The current test loops over the result directory versions, and while
it aggregates and prints the individual failures, the unittest
framework has the same purpose.

To me, as a user, I'd rather know that test "xyz" failed right away,
than having to parse the output of a longer text message.

Signed-off-by: Cleber Rosa <crosa@redhat.com>