Python 3: make the subcommands sub parser a required option
The goal of this change is to make the behavior of the Avocado command
line application as similar as possible to the behavior when running
under Python 2.  This bit is about what happens when the simplest
possible Avocado command line is run:

 $ avocado

On Python 2, it prints the complete usage, along with the error
message "too few arguments".  While we could go employ extreme changes
to make the very same error message be given on Python 3, this seems
wrong.  The reason is that the message is generated by the Python 2
standard library.  The Python 3 standard library relies on "required"
arguments instead.  When settings the subcommand sub parser as a
required option, the behavior is exactly the same, minus the error
message which becomes "the following arguments are required:
subcommand".

Signed-off-by: Cleber Rosa <crosa@redhat.com>