Python 3 port: address avocado.utils.astring.string_to_safe_path
The main goal here is to make string_to_safe_path behave properly
under Python 3.  The big change here is the added code that selects
the module where the maketrans() function is located, which in Python
3 is a static method of bytes.

Additionally, it removes the module level definitions
(FS_UNSAFE_CHARS, FS_TRANSLATE) which don't have any other use so far.

FS_USNAFE_CHARS has become a simple variable, now expressed as a
sequence of bytes.  If other characters are added to the translation
table, maketrans() will make sure the they are the same size, so
I believe the code to generate the second argument of the same size
can be dropped.

Now, a side effect is that, on Python 2, the result of the json
parsing (json.loads()) is a dict with unicode keys and values.  When
that is passed to the updated string_to_safe_path() implementation, a
causes TypeError is raised (from Objects/unicodeobj.c,
charmaptranslate_lookup()).

Signed-off-by: Cleber Rosa <crosa@redhat.com>