Python 3 port: address avocado.utils.astring.string_safe_encode
And its related functions (callers and callees).

In Python 3, strings are unicode, so an encode operation will
return a byte object, while a decode (on a byte or bytearray)
will return a string (unicode).

To start with, string_safe_encode() wants to solve the problem
of users mising unicode and other encoded strings, which doesn't
make sense for Python 3.

Additionally, it uses encode and decode in a way that is
very specific to Python 2, such as assuming the result of a
decode() can then be "encode()d" and still be a string.

For Python 3, the idea here is to *not* deal with any
encode/decode operations attempting to get a "safe" string.
The first part of the function, which converts non-string
objects to string, is still being relied upon by some code,
so it's kept (at least for now).

Signed-off-by: Cleber Rosa <crosa@redhat.com>