utils.astring: Bring back FS_UNSAFE_CHARS and TRANSLATE
The FS_UNSAFE_CHARS could be useful to users when dealing with filenames
but not in need of the full "string_to_safe_path" function.

The FS_TRANSLATE was also brought back, but it's usability outside this
module is questioned, therefor it's created as private variable. The
reason to include this on module level is to only create the translation
once and not per each call.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>