output: Adjust LoggingFile to work with raw output
The 4235b3299210c24f91539754312c8ad09053cf32 changed the
stdout/stderr/output logging handlers to not to add tailing '\n's, but
these are processed by LoggingFile, which is designed to remove the
'\n's and to logged them per line. As a result the unittest was
incorrectly changed as the first outputs actually add the '\n' which was
removed in the expected output.

This patch keeps the raw approach as it is necessary for the last
buffered line and instead of buffering puts all lines including the
tailing '\n' and not '\n' for the remaining last line (if it's
available). Consequently this should improve the situation where output
is being produced by stderr and stdout concurrently as no buffering of
the non-tailing-newline output is done on LoggingFile level.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>