avocado/core/loader.py: simplify and improve detection of Python files
The use of PathInspector checks too much and too little at the same time.
For once, it attempts to find Python code inside a file that doesn't have
a '.py' suffix.

This means that a file named "test" instead of "test.py" can be recognized
as an INSTRUMENTED test.  A file no ending in ".py" can not be currently
loaded as a Python module by Avocado, and it results in a test execution
ERROR.

This changes the behavior for files with Python code, but not ending
in '.py' that can not be loaded as INSTRUMENTED tests, and will now
result in NOT_A_TEST instead.

Signed-off-by: Cleber Rosa <crosa@redhat.com>