Python 3 port: drop 2to3 use from setup.py
The 2to3 feature was enable to give Avocado a head start in the port to
Python 3.  But, while having positive characteristics, it has some
drawbacks, specially at development time, in which an extra step is
required.

The porting work done so far has not relied on 2to3, and it's been
having good results.  Let's then drop 2to3 completely, and allow the
same code base, without any massaging, to be used on both Python 2 and
Python 3.

Signed-off-by: Cleber Rosa <crosa@redhat.com>