Python 3 port: use AvocadoParams.iteritems() method
As part of 0795831, all occurrences of dict.iteritems() turned into
iteritems(dict).  The problem is that AvocadoParams doesn't implement
the items() method that six.iteritems() expects from Python 3.

So given that AvocadoParams doesn't provide it, we could either implement
items() just to satisfy iteritems(), or just use the existing interface.

This commit chooses the later, which means reverting a small part of
the changes (incorrect ones) from 0795831.

Signed-off-by: Cleber Rosa <crosa@redhat.com>