loader: defaults to recursive
We currently have to tag a sub-class from a class inheriting from
avocado.Test to make sure it will be considered an Avocado test class.
The recursive behaviour is proven to work by now, so it's time to make it
the default behaviour.

This patch removes the need of a "recursive" docstring, making the
recursive discovery the default behaviour. The "enable" docstring
was also removed for the same reason.

Signed-off-by: Amador Pahim <apahim@redhat.com>