FDDrainer: flush the file based handlers before filling results
The FDDrainer reads from the PIPEs and writes to both the internal
StringIO instances and to stream logging handlers.  There are no
guarantees that the logging handlers will flush the content before the
FDDrainer finishes.

Let's explicitly close all the handlers (which should really be
FileHandlers) associated with the stream loggers, which should flush
and sync the content on the files themselves.

This *should* fix the race conditions with content missing from the
output files that we're observing on Travis-CI.

Signed-off-by: Cleber Rosa <crosa@redhat.com>