FDDrainer: flush the file based handlers before filling results
The FDDrainer reads from the PIPEs and writes to both the internal
StringIO instances and to stream logging handlers.  There are no
guarantees that the logging handlers will flush the content before the
FDDrainer finishes.

Let's explicitly close all the handlers (which should really be
FileHandlers) associated with the stream loggers, which should flush
and sync the content on the files themselves.

Additionally, the combined drainer was being missed from the flush
process, and was causing some output in that mode to be unavailable on
generated output.

Signed-off-by: Cleber Rosa <crosa@redhat.com>