avocado/utils/script.py: add a read only file permission mode shortcut
The script module is frequently used in Avocado itself in its
functional tests to generate tests files that will be run by the
`avocado` command line test runner.

Being able to create files that are *not* executable is important
because it can determine if a file will be identified as a SIMPLE or
INSTRUMENTED test.  Let's make that read only file permission
available as a definition in the module level, just like
`DEFAULT_MODE` is there.

Signed-off-by: Cleber Rosa <crosa@redhat.com>