Output check: simplify logic of check and record output
First, the recording of the output should be governed by the
corresponding setting alone.  There's no need to check the test being
run in job standalone mode, because if that's being run like that,
there's no way to actually set the record mode.

Then, for the output check mode, it won't happen if the the test is
being run in standalone mode or if the output recording has been
explicitly disabled (then we have no material to check the reference
with) or if the checking itself has not been enabled.

Signed-off-by: Cleber Rosa <crosa@redhat.com>