Remove 36.0 (LTS) to 51.0 (pre 52.0 LTS) jobdata support
And include jobdata results and a test for 52.0 (LTS).

The 52.0 release, an LTS release, was meant to allow transitions from
the previous LTS version (36.0lts) and it did so including the ability
to replay jobs created with the older version.

Now, we're 6+ months into the 52.0 release, and as promised, we'll
stop maintaining 36.0lts.  To keep to code clean, let's remove all
this compatibility with versions before 52.0.

Signed-off-by: Cleber Rosa <crosa@redhat.com>