Parameters: remove knowledge about a specific (job) logger
The forward reference from the parameter module towards the job log
is not ideal.  The parameter module should be used by test (and
related code) and should not know about the test and job internals.

But, there's the clear requirement and convenience of logging the
access to the parameters.  This makes the user of the parameter
instance define which logger should be used, and not the other
way around.

Logger names are used here, because, the communication between
runner and test (get_state()) currently includes test parameters,
and serializing a logger instance is not possible.

Signed-off-by: Cleber Rosa <crosa@redhat.com>