Test/Runner: make the test state content slimmer and documented
While Avocado is running a test, it constantly requests the test to
send its state, making this operation a good candidate for some
optmization.

This change reduces the amount of attributes that are accessed and
serialized and returned as a test state to the barely minimum
observed, and at the same time, it documents it so that it will
hopefully start to serve as a documentation for the (private) protocol
that runners and tests use.

While at it, as a another smaller optmization, this list is defined at
the module level as a tuple (order doesn't matter here) and prevents
the list from being created on every single state construction.

Signed-off-by: Cleber Rosa <crosa@redhat.com>