avocado.utils: Improve path in "software_manager.get_source"
Currently it uses "--urls" to list urls to download from and then it
initiates another execution to get the actual file. This is potentially
dangerous as the name can change between those two executions. Let's get
the location from the last line of the download command.

Note: This requires some fixes to process/subprocess as the last line is
usually not displayed.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>