utils.software_manager: Improve the way src.rpm files are downloaded
Currently it uses "--urls" to list urls to download from and then it
initiates another execution to get the actual file. This is potentially
dangerous as the name can change between those two executions.

Due to yumdownloader bug it's not possible to get the downloaded file
name from python execution (as in non-TTY mode it's not printing it on
<F26) let's use tmpdir and assume the only existing "src.rpm" there is
the freshly downloaded one.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>