selftests/checkall: enable parallel run for plugin tests
This change enables plugins tests to also be run in parallel, if
they exist.

Because the "selftests/{unit,functional,doc}" directories are valid
Python modules, that is, they contain `__init__.py` files and can
be imported by the unittest runner, no special handling is necessary.

But, since the complete path for directories which (can) hold plugins
tests are not valid Python modules (and shouldn't be), some special
handling of directories is necessary.

Because of that, it's not currently possible to run *all* of the
tests (from "core" selftests and from plugins) at the same time.
A future enhacement would be to better schedule processes that
can be loaded from the same path at the same time, but all of
them at once with global "test slices".

Signed-off-by: Cleber Rosa <crosa@redhat.com>