Tree: add unittests
The `avocado.core.tree.TreeNode` is the foundation of the current
Parameter System (`avocado.core.parameters.AvocadoParams`), but it
lacks unittests that only deal with itself.

Some tests that exercise the Multiplexer version of the `TreeNodes`
(`MuxTreeNodes`) exist, but again, that's not what the core Parameter
System is based on.

Signed-off-by: Cleber Rosa <crosa@redhat.com>