Travis-CI: use documented syntax for virtualenv + system_site_packages
According to https://docs.travis-ci.com/user/languages/python, the
standard way of setting up that configuration is by using the
"system_site_packages" under "virtualenv".

It looks like the current syntax was a valid fix for a Travis-CI
issue that is now fixed, and it was being used in Avocado to support
the now dropped Python 2.6 version checks.

Signed-off-by: Cleber Rosa <crosa@redhat.com>