HTML plugin: use subprocess.DEVNULL instead of custom open file
The HTML plugin is using the `subprocess` library to fire up the
web browser that will show the report.  In theory, we could be
using our own `avocado.utils.process` for that, but given the
comments in that section, I believe there may be caveats.

As a fix for the immediate issue at hand (pylint catching a failure on
a Python 3 environment about using `file`), let's just use the
definitions from the `subprocess` module itself.

Signed-off-by: Cleber Rosa <crosa@redhat.com>