Python 3 port: work around fact that MuxTreeNode is not hasheable
On Python 3, many other objects are not natively hasheable, including
our own `avocado_varianter_yaml_to_mux.mux.MuxTreeNode`.  When checking
for the number of counts in `avocado.core.parameter.get_or_die()`, a
set() is used, which implicitly calls `hash()` on a `MuxTreeNode`.

This won't work on a Python 3 environment:

 >>> import avocado_varianter_yaml_to_mux.mux
 >>> hash(avocado_varianter_yaml_to_mux.mux.MuxTreeNode())

Let's use an identity check instead for the number of results we
found, and avoid that incompatibility.

Signed-off-by: Cleber Rosa <crosa@redhat.com>