Python 3 port: use input() instead of raw_input()
On Python 2, input() has a different behavior, including an evaluation
of the obtained string.  On Python 3, it just returns the stripped
string.

Let's hand off the difference handling to six.

PS: we won't use use six.moves.input() on the failtest_ugly.py file,
given that it's intended to be a bad example.

Signed-off-by: Cleber Rosa <crosa@redhat.com>