avocado/core/loader.py: implement ordered test discovery
The order in which test classes and methods are found is relevant
to the loader result.  Currently, this ordering is not dependable
as the loader findings are stored on a regular dictionary.

This change makes the ordering dependable, and so happens to also
make the test pass on a Python 3 environment.

Signed-off-by: Cleber Rosa <crosa@redhat.com>