Settings: remove variables from the main module namespace
The main module namespace has a lot of variables which are only
used in the main class initialization method.  Let's avoid poluting
the module namespace and move those to where they're used.  If
we want to put variables at the global module namespace, we
should consider if they're useful to other modules, make them
all caps, and document them.

A nicer refactor could be done here, but IMO, this is still an
improvement.

Signed-off-by: Cleber Rosa <crosa@redhat.com>