Packaging: add bash RPM sub package
To remove some of the bloat (even if little) of the base
python-avocado package, let's move content that is not useful to
Python developers out of the way.

This should also make the experience of content more similar across
users of the Python module (say, from a `pip install`) and users of
the python-avocado package.

Signed-off-by: Cleber Rosa <crosa@redhat.com>