Packaging: make config files data files of the avocado module
In a pure Python world[1] the "avocado" module requires the existence
of these configuration files.  Let's make those configuration files
"data files" of that module.

Since the configuration location can now be a different, the settings
module needs adjustments to also attempt to load them from package
resources (which is tied to the data files).

Finally, when using RPM packages, we need to put the configuration
back into the system location (/etc).

[1] - that is, ignoring operating system or distribution specific
      conventions and packaging technologies

Signed-off-by: Cleber Rosa <crosa@redhat.com>