utils.cpu: Update cpu info pattern for aarch64
arm64 or aarch64 pattern is not easy to identify in the
cpuinfo. For v7 there is model name in the cpuinfo and it
could be regonize as arm for 32 bit. But in v8 only could
found 'CPU architecture' in the cpuinfo and current code
will fail to match and return i386 which is wrong.

For fix this, if 'CPU architecture' found in cpuinfo, use
platform.machine to return the arch name.

Signed-off-by: Wayne Sun <gsun@redhat.com>