libexec: Move libexec to package_files
With the recent changes the "libexec" dir is only available when
installed via RPM. In order to have "libexec" files available
as standard package we have to include it in  the package_files
dir (avocado/). The benefit is that we can actually use package
resources to get the location in "avocado exec-path" command.

For packaging purposes let's simply symlink $libexec/avocado to the
packaged $avocado/libexec location so there is the usual location, but
keep using the python-location by default.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>