settings: Change the config-file order
Recently we introduced "in-package-config" which is packaged by default
with Avocado and in the original version if exists the system-wide
config is not used at all. I think this is unexpected behavior as users
are used to set configs in /etc so this patch changes it to first use
in-package-config, then in-etc-config and last the user-config.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>