avocado.utils.process: use bytes for raw stdout/stderr
This brings a change in behaviour, in which the stdout/stderr of the
executed process will now be a stream of bytes, instead of text
(string/unicode type).

A new attribute is added, in which the same content is available as
text, decoded on the fly.  To *not* have to deal (and keep) with
multiple copies of the same data (bytes and bytes decoded to text) a
property is used.  This solution chooses to use more CPU time over
more memory.  This lets to the distcretion of the user, to keep a copy
of the data (decoded as text) if he/she sees that as a better
tradeoff.

Signed-off-by: Cleber Rosa <crosa@redhat.com>