Test statuses: do not rely on varianter_yaml_to_mux plugin
Which is an optional plugin, and the test statuses feature is at the
innermost core of Avocado.  While the original test indeed saved a lot
of duplicated code, it's code that serves a specific testing purpose,
and it's better to be simpler and more verbose and with the right (no)
dependencies.

This issue came up during the Python 3 port work.  Now that all the
unittests (selftests/unit/*) were passing (locally, there's one other
PR pending to be accepted), the next logical step is getting the
"core" functional (selftests/functional/*) tests passing.  Because
the 'varianter_yaml_to_mux' optional plugins port is not done, the
"core" functional tests fail and the current work flow is broken.

Signed-off-by: Cleber Rosa <crosa@redhat.com>