Test statuses: do not rely on yaml_to_mux varianter plugin
Which is an optional plugin, and the test statuses feature is at the
innermost core of Avocado.  While the original test indeed saved a lot
of duplicated code, it's code that serves a specific testing purpose,
and it's better to be simpler and more verbose and with the right (no)
dependencies.

Signed-off-by: Cleber Rosa <crosa@redhat.com>