dispatcher: Unify the map_method implementations
Currently almost each "Dispatcher" implements it's own "map_method",
some allow to log tracebacks, some doesn't... Let's implement 2 main
variants, one simple without reporting the results, the other with
support for deepcopying of the arguments and returning the list of
results and use them in all dispatchers.

The main difference against previous version is that all dispatchers
will log the traceback (to avocado.debug) and that instead of individual
arguments it supports *args.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>