Runner: let multiprocessing module choose the context type
It was noticed during a review that, even though the SimpleQueue
requires a context on Python 3, the multiprocessing module can choose
one itself.

Since I originally had no intention to force the `spawn` type of
context, let's use that instead.  Also, since we're using different
names for a helper function and the class, one of the inspekt/pylint
exceptions is not needed anymore.

Kudos to Amador Pahim for suggesting that.

Reference: https://github.com/avocado-framework/avocado/pull/2441#discussion_r167850053
Signed-off-by: Cleber Rosa <crosa@redhat.com>