avocado.utils.iso9660: always deal with bytes
The avocado.utils.process module has been changed so that the content
of stdout and stderr are now bytes.  The content generated by the
iso9660 backend implementations should also operate on bytes, given
that it makes little sense for the library to assume a given content
type of a file (inside an iso9660 archive).

To make the data returned by the various backend `read()`
implementations always return bytes on both Python 2 and 3, an
explicit conversion is done from strings to bytes (which affects only
Python 2, since Python 3 already operates with bytes on files opened
in binary mode).

Signed-off-by: Cleber Rosa <crosa@redhat.com>